Move Updateguide to service
